% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestModel.R
\name{getBestModel}
\alias{getBestModel}
\title{getBestModel
This function provides a dataframe with the list of models of nucleotide evolution sorted by AICc (from lowest to highest)}
\usage{
getBestModel(y)
}
\arguments{
\item{y}{takes as input a dataset of aligned sequences of the same length and known ORF}
}
\value{
returns a dataframe with the list of models of evolution sorted by AICc, from the lowest to the highest AICc value
}
\description{
getBestModel
This function provides a dataframe with the list of models of nucleotide evolution sorted by AICc (from lowest to highest)
}
