% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRF.R
\name{getRF}
\alias{getRF}
\title{getRF
This function determines the Reading Frame (RF) of a dataset of nucleotide sequences}
\usage{
getRF(alignedFasta)
}
\arguments{
\item{alignedFasta}{takes as input a dataset of aligned sequences of the same length

As output, this function either renames the input file if the sequences are in the ORF 1, or it writes a new fasta file if the input file was in ORF 2 or 3.
The output file will be of RF 1}
}
\description{
getRF
This function determines the Reading Frame (RF) of a dataset of nucleotide sequences
}
