% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingBestModel.R
\name{gettingBestModel}
\alias{gettingBestModel}
\title{gettingBestModel
This function provides a dataframe with the list of models of nucleotide evolution sorted by AICc (from lowest to highest)}
\usage{
gettingBestModel(fastaORF)
}
\arguments{
\item{fastaORF}{takes as input a dataset of aligned sequences of the same length and known ORF}
}
\value{
returns a dataframe with the list of models of evolution sorted by AICc, from the lowest to the highest AICc value
}
\description{
gettingBestModel
This function provides a dataframe with the list of models of nucleotide evolution sorted by AICc (from lowest to highest)
}
